Moved handlers into single file, cleanup
