+++ client/src/components/LeafletMap.js
@@ -1,9 +1,8 @@
 import React, {useState} from "react";
 import {MapContainer, TileLayer} from "react-leaflet";
 import MarkerClusterGroup from "react-leaflet-markercluster";
-import InBoundersMarkups from "./LeafletLocalData"; // Local data
-import ZoomHandler from "./LeafletZoomHandler"; // Zoom handler
-import CenterLocatorHandler from "./LeafletCenterViewHandler"; // Locator handler
+import InBoundersMarkups from "./LeafletDataDB"; // Requests to server
+import {CenterLocatorHandler, ZoomHandler} from "./LeafletHandlers"; // Zoom and locator handler
 import {
     initialCenterCoordinates,
     mapLimits,
@@ -62,35 +61,28 @@ const Map = props => {
                 }
                 url={"https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"}
             />
-            {
-                /* zoomLevel > 16 ? ( */
-                <MarkerClusterGroup
-                    /**
-                     * Documentation markercluster
-                     * https://www.npmjs.com/package/leaflet.markercluster
-                     * https://github.com/Leaflet/Leaflet.markercluster#usage
-                     */
-                    disableClusteringAtZoom={19} // Disable clsuetering and display all individual markers
-                    maxClusterRadius={100} // Set number of cluster around eachother. Bigger value mean less cluster markers.
-                    showCoverageOnHover={false} // Show area of the cluster
-                    chunkedLoading={true} // Boolean to split the addLayers processing in to small intervals so that the page *may* not freeze.
-                    spiderfyOnMaxZoom={false}>
-                    {/* Show markers linked when close to eachother */}
-                    <InBoundersMarkups
-                        zoomLevel={zoomLevel}
-                        Bounders={boundsView}
-                        userName={userName}
-                        userTrees={props.userTrees}
-                        setUserTrees={props.setUserTrees}
-                        userLeaves={props.userLeaves}
-                        setUserLeaves={props.setUserLeaves}
-                    />
-                </MarkerClusterGroup> /* 
-            ) : (
-                <MaxZoomClusters />
-            )} */
-            }
-            {/* <Rectangle bounds={mapLimits} /> // Rectangle to display map limits*/}
+            <MarkerClusterGroup
+                /**
+                 * Documentation markercluster
+                 * https://www.npmjs.com/package/leaflet.markercluster
+                 * https://github.com/Leaflet/Leaflet.markercluster#usage
+                 */
+                disableClusteringAtZoom={19} // Disable clsuetering and display all individual markers
+                maxClusterRadius={100} // Set number of cluster around eachother. Bigger value mean less cluster markers.
+                showCoverageOnHover={false} // Show area of the cluster
+                chunkedLoading={true} // Boolean to split the addLayers processing in to small intervals so that the page *may* not freeze.
+                spiderfyOnMaxZoom={false}>
+                {/* Show markers linked when close to eachother */}
+                <InBoundersMarkups
+                    zoomLevel={zoomLevel}
+                    Bounders={boundsView}
+                    userName={userName}
+                    userTrees={props.userTrees}
+                    setUserTrees={props.setUserTrees}
+                    userLeaves={props.userLeaves}
+                    setUserLeaves={props.setUserLeaves}
+                />
+            </MarkerClusterGroup>
         </MapContainer>
     );
 };
